java.flattenStreams: correctly handle flattening metafactories using H_NEWINVOKESPECIAL
these require that the new instance be created BEFORE the arguments are pushed onto the stack, so i've had to split up the handle flattening methods in BytecodeHelper into a pre and post phase.

also added improved logging for stream flattening to help disambiguate the stream call chains being transformed.